+++ TransferEasy.Domain/ApplicationService.cs
@@ -2,29 +2,29 @@
 
 public interface IApplicationService
 {
-    void Transfer(double amount, int fromAccountId, int toAccountId);
-    void Deposit(double amount, int toAccountId);
-    void Withdraw(double amount, int fromAccountId);
+    void Transfer(decimal amount, int fromAccountId, int toAccountId);
+    void Deposit(decimal amount, int toAccountId);
+    void Withdraw(decimal amount, int fromAccountId);
     IEnumerable<Account> GetCustomerAccounts();
     IEnumerable<Account> GetSystemAccounts();
     Account GetAccount(int accountId);
     int CreateAccount(string username);
-    double GetAccountBalance(int accountId);
+    decimal GetAccountBalance(int accountId);
 }
 
 public class ApplicationService(IProvideBalance balanceProvider, IPublisher commandPublisher, IAccountService accountService) : IApplicationService
 {
-    public void Deposit(double amount, int toAccountId)
+    public void Deposit(decimal amount, int toAccountId)
     {
         commandPublisher.Publish(new DepositCommand(toAccountId, amount));
     }
 
-    public void Transfer(double amount, int fromAccountId, int toAccountId)
+    public void Transfer(decimal amount, int fromAccountId, int toAccountId)
     {
         commandPublisher.Publish(new TransferCommand(fromAccountId, toAccountId, amount));
     }
 
-    public void Withdraw(double amount, int fromAccountId)
+    public void Withdraw(decimal amount, int fromAccountId)
     {
         commandPublisher.Publish(new WithdrawCommand(fromAccountId, amount));
     }
@@ -42,7 +42,7 @@ public class ApplicationService(IProvideBalance balanceProvider, IPublisher comm
         return accountService.GetAccount(accountId);
     }
 
-    public double GetAccountBalance(int accountId)
+    public decimal GetAccountBalance(int accountId)
     {
         return balanceProvider.GetBalance(accountId);
     }
