+++ TransferEasy.Domain/Commands.cs
@@ -1,6 +1,6 @@
 ﻿namespace TransferEasy.Domain
 {
-    public record TransferCommand(int OriginAccount, int DestinationAccount, double Amount);
-    public record DepositCommand(int DestinationAccount, double Amount);
-    public record WithdrawCommand(int OriginAccount, double Amount);
+    public record TransferCommand(int OriginAccount, int DestinationAccount, decimal Amount);
+    public record DepositCommand(int DestinationAccount, decimal Amount);
+    public record WithdrawCommand(int OriginAccount, decimal Amount);
 }
